[Refactor] Repository countFestivalsByUserIdAndStatus문을 leftJoin으로 끌고 와서 Count하기

+++ src/main/java/com/amp/domain/organizer/service/OrganizerMypageService.java
@@ -32,13 +32,13 @@ public class OrganizerMypageService {
                 .orElseThrow(() -> new CustomException(OrganizerErrorCode.ORGANIZER_NOT_FOUND));
 
         // 진행 중인 공연 수 조회
-        Integer ongoingCount = organizerRepository.countFestivalsByUserIdAndStatus(
+        Long ongoingCount = organizerRepository.countFestivalsByUserIdAndStatus(
                 userId,
                 FestivalStatus.ONGOING
         );
 
         // 진행 예정 공연 수 조회
-        Integer upcomingCount = organizerRepository.countFestivalsByUserIdAndStatus(
+        Long upcomingCount = organizerRepository.countFestivalsByUserIdAndStatus(
                 userId,
                 FestivalStatus.UPCOMING
         );
